Give up on httpbin access from appveyor
It appears that these workers are more or less constantly blocked. There
is little point in all these false-positive failures.

The respective tests will continue to run on appveyor, but limited to
deployments with a dedicated docker-instance of httpbin running as
part of the CI setup (this is ubuntu for now).