Fix EnsureHashAlgorithm & use EnsureListOf
The EnsureHash algorithm now calls its super.__init__ correctly.

We use EnsureHashAlgorithm() | EnsureListOf(...) in ls_file_colelction
now. This correctly handles incoming values, which are always a list
in CLI (argparse behaviour with --append) and can be str or list in
Python API. At least when they are OK.

When input is not OK, two things are suboptimal:
- an incorrect hash (str) proceeds to EnsureListOf, which tries to go
  letter by letter, failing on first
- a hash (list) with an incorrect value would first fail on
  EnsureHashAlgorithm(list), proceed to EnsureListOf, and fail on
  EnsureHashAlgorithm(str)
As a conequence, in both cases the error message would say "does not
match any of 2 alternatives" and print EnsureChoice message ("is not
one of ...") twice.