calculate download size if tell is not available
This commit modifies _stream_download_from_request
to calculate the current download size from the
length of the downloaded chunks instead of using
`tell()`, if the HTTP-response has no
`content-length` header. That ensures that
progress is reported.

TODO: we should determine why a non-raising
http-request for a file would not contain a
content-length header. It may be 201-response,
but I could net yet recreate the condition.