Enhance `FileSystemItem` and `iter_tar()` to provide file IO
This enabled something like this:

```
In [1]: from pydicom import dcmread
In [2]: from datalad_next.iter_collections.tarfile import iter_tar
In [3]: for i in iter_tar('/tmp/icfstore/dl-Z03/P000624_dicom.tar'):
   ...:     if i.fp:
   ...:         with dcmread(i.fp) as dcm:
   ...:             ...
```

Or in other words, arbitrary (post-)processing of files in a collection.

For now this is just for `iter_tar()`, but in a next step it can be done
for all iterators. When done, the multi-hash logic can move out of the
iterators and into `ls_file_collections`.