BF: cache archivist handler for later reuse
In INM-ICF-utilities, a get operations on a directory with many files from the
same tarball failed after the first file was successfully retrieved.
Debugging revealed that the internal retrieval of the tar archive was attempted
for this first try successfully. Later files retried retrieval unnecessarily,
resulting in cryptic errors downstream