reduce duplication in credential prompt
Co-authored-by: Michael Hanke <michael.hanke@gmail.com>

This commit reduces duplication in the
credential prompt that is used for
401-HTTP server replies. This is done
by just presenting the authentication
realm value without the leading
URL-scheme and URL-location part. The
latter two can be derived from the
protected URL, that is also shown in
the credential prompt.

Before the changes in this branch the prompt
to access, for example, 'docs.unm7.de/learning' was:

    Credential needed for access to https://docs.inm7.de/learning

With the changes in this branch the prompt is now:

    Credential needed for accessing https://docs.inm7.de/learning (authentication realm 'https://docs.inm7.de/INM-7 Documentation')