Show realm in credential prompt (#396)
This change inserts the realm into the
credential prompt for 401-error handling.

Because the realm is a just a name for a
set of resources, it is indenpendent from
the URL-path. That means that the URL-path
cannot be used to determine the realm.

For example, the test fixture
`http_server_with_basicauth` creates a
server whose resources are all in the
protection realm 'Protected', independent
of the URL-path, which is typically:
`/testfile.txt`. Another example is the
URL `https://docs.inm7.de/`. All paths
in this host, e.g.
`https://docs.inm7.de/learning` are
protected by a realm with the name
`INM-7 Documentation`.

Adding the realm to the prompt, allows
a user to manually enter credentials
for an authentication realm via
`datalad credentials set`.

The prompt is tailored to reduce redundancies
by just presenting the authentication
realm value without the leading
URL-scheme and URL-location part. The
latter two can be derived from the
protected URL, that is also shown in
the credential prompt.

Before the changes in this branch the prompt
to access, for example, 'docs.unm7.de/learning' was:

    Credential needed for access to https://docs.inm7.de/learning

With the changes in this branch the prompt is now:

    Credential needed for accessing https://docs.inm7.de/learning (authentication realm 'https://docs.inm7.de/INM-7 Documentation')