change return type of ArchiveOperations.open()
This commit changes the return type of the
method `open()` of `ArchiveOperations` and
its subclasses to reflect the return type
the is defined in the method code. That means
the return type is changed to `Generator`.

`Generator` is not the return type that is
observed by users of instances of, for
example, `TarArchiveOperations`. This is
due to the `@contextmanager`-decorator on
`open()`. Therefore this commit adds a
doc-string to `TarArchiveOperations.open()`
that describes the behavior that a caller
on `open()` on an instance of
`TarArchiveOperations` would observe.