use PurePosixPath for names of zipfile-entries
This commit modifies zip-iterators to use
`PurePosixPath` instead of `PurePath`. The
reason is that, according to zip-APPNOTE.TXT
version 6.3.3 from 2012-09-01, all slashes in
file names (contained in zip) should be
forward slashes, i.e. '/' (see
https://pkware.cachefly.net/webdocs/APPNOTE/APPNOTE-6.3.3.TXT)
zip-APPNOTE.TXT from 2006-09-26 states that
all slashes should be forward slashes.

It seem to be safe to assume that almost all
zip-files we encounter will be using forward
slashes and using a `PurePath` on Windows
to represent names of files in a zip-archive
will not reflect the content properly. In
addition, those paths cannot be used to
retrieve items from a `zipfile.ZipFile`
without conversion to a posix-style path.

We should therefore use `PurePosixPath`-instances
to represent the name of elements in a
zip-archive.