use '/' in zip-file member names in test
This commit ensures that all zip-file names
that are used as parameters in methods of
ZipArchiveOperations use '/' in names
of zip-archive items.

That is due to the fact that zip enforces
the use '/' in the names of its items
(zip-APPNOTE.TXT 6.3.3 from September 1st,
2012).

Until PR https://github.com/datalad/datalad-next/pull/409
is merged this will ensure that the proper
file names are used to check for items in
zip-archives. The commit can be reverted
once PR  https://github.com/datalad/datalad-next/pull/409
is merged.