Recognize datalad-osf legacy deposit and unpack them properly too
This makes it possible to use this code *instead* of the old git-remote
helper in `datalad-osf`.

This is intentionally implemented as a on-read-only feature. This
enables cloning from such deposit (forever). However, a push will
replace the deposit with the new/standard layout, and from there on
require a modern `datalad-osf` or access via `datalad-annex::` directly.