Handle HEAD branch conflicts in `datalad-annex::`
Whenever a repo had a HEAD branch that differed from the configured
default branch in Git, the mirror repo would be on the configured
default. On push this configured default would become the configured
HEAD on the remote, and on clone this would fail to yield a
checkout, because the actual repo content might not even
have that branch.

A typical use case would be a repo with a `main` branch, but a system
that has `master` to be the configured default.

Closes #412