Make `TarfileItem.name` be of type `PurePosixPath`
Rational from https://github.com/datalad/datalad-next/pull/409:

```py
>>> PureWindowsPath(*PurePosixPath('d/a\\b\\c.txt').parts)
PureWindowsPath('d/a/b/c.txt')
```

This means that we must relay the POSIX nature of the archive member
path to the users, because there is no way to express this as a platform
(windows) path -- and also no way to extract this file under an
equivalent name on an FS that uses windows-semantics. So a type
mismatch can be used to trigger mitigation strategies.

If feel like a clean(er) solution would be to change `TarFileItem` to
declare to have a `name` of type `PurePosixPath`.

For the same reason and rational, a symlink target must also be
communicated in POSIX form.