declare ZipfileItem.name type as PurePosixPath
This commit changes the type of ZipfileItem.name
to `PurePosixPath`. This is to ensure that instances
will properly represent zipfile-items with
challenging names, e.g. backslash in a name,
correctly on all platforms.

This has become necessary because the preivously
used `PurePath`-type would yield `PureWindowsPath`-
instances on Windows. Those could not properly
handle zipfile-item names that contain backslashes
(because `PureWindowsPath` interprets backslashes
as directory name-separators and would therfore
interpret a name like "a\b\c" as file "c" in the
directory "a/b".

The commit also includes a zipfile-item with a
name that is not a valid Windows-path name in the
test zip-file. (The referred to name does not
contain a backslash, because on Windows the
backslash does not make it into the zip-file item
name.)