add a class to represent dir-names in zip-archives
This commit adds the class `ZipFileDirPath`. The
class is a direct subclass of `PurePosixPath`.
Instances of the class represent directory names
in zip-archives. The difference to `PurePosixPath`
is that rendering appends a '/' to the name.

This allows us to use the generic way of testing
for containment in ZipArchiveOperations, i.e.

  if item.name in archive:

Otherwise we would have to discriminate
between directory- and file-names and manually
add a '/' to directory names, e.g. something
like this:

  if (item.name + '/' if item.type == FileSystemItemType.directory else item.name) in archive: