add a batch command context handler with tests
Co-authored-by: Michael Hanke <michael.hanke@gmail.com>

This commit adds a batch-command-context handler that
gives the same termination strategies as the
run-context handler.

The tests contain an example for a protocol that
is capable of interpreting responses of not
a-priori known length or structure (python
program output). The protocol is
`PythonProtocol`, which is used  in
`test_plain_batch_python_multiline`.