add data processor pipeline driver and processors
This commit adds a driver for a pipeline of data
processors that can convert chunks of bytes into
"derived" data, e.g. string, JSON objects, etc.

The driver can be used imperatively by calling
two methods on a driver object.  It can also be
used as a generator, fetching bytes from a
base generator and yielding the output of the
last data processor in its pipeline.

In addition this commit contains four data
processors:

- a decoder
- a jsonline-reader and loader
- a pattern aligning processor
- a processor that splits bytes or string int
  lines