Let the test battery honor DATALAD_TESTS_NONETWORK
This is primarily achieved via a custom pytest mark. However, there is a
session-level fixture that itself depends on network access. Unlike with
a function-level fixture, I cannot figure out how to make any dependent
test be skipped. Instead, I need to use the mark as a skipping
decorator.

With these changes I can successfully run the test suite on a system
with no network access (not just a proxy-based diversion):

```
247 passed, 39 skipped, 19 warnings in 171.10s (0:02:51)
```

Ping #490