keep protocol keyword args separate
This commit addresses comment
https://github.com/datalad/datalad-next/pull/484#discussion_r1368172700
about cluttering the namespace.

The commit introduces a convention
that is currently only used in the
protocol class `KillWrapper`. The
single keyword argument of the
class is prefixed with `dl_`. We
should claim the name space `dl_*`
for keyword arguments in the
protocols that we define.

This commit also fixes the return
value doc-string description of
`_create_kill_wrapper`.