add pipeline processing generator-protocols
This commit adds pipeline processing generator
protocols. The protocols can be initialized
with a list of processors that they will apply
to the data that they receive.

There is currently no equivalent for
non-generator protocols. The reason for that
is that those protocols expect every data
element that they handle to be `bytes`.
That does not fit well with processors
like `splitlines_processor` or
`decode_processor`.

We could could implement
a new non-generator protocol "family"
that collects whatever comes from a
processor pipeline as the values of
`stdout`- and `stderr`-keys of the
result dictionary. But it is currently
not clear when and where that would
be used.