reorder data processor code, update doc-strings
This commit addresses comment:
https://github.com/datalad/datalad-next/pull/484#discussion_r1369655832
It removes unrendered doc strings and
reorders data processor code so that
the data processor is on top of the
source files. In addtion it adds
`__all__` variables to limit imports
to the objects that constitute the
user-interface.

The commit also adds doc-string
rendering for the module
`datalad_next.runners.data_processors`
This includes markup fixes