introduce a KillingGenerator-wrapper
This commit introduces a wrapper for a result
generator object that will trigger timeout
handling whenever the generator is iterated over.

That will limit a single iteration on the result
generator to the timeout that was specified.

Because we use the KillGenerator wrapper now in
`run()` to equip the result generator with the
arming/disarming capabilities, `BatchProcess`
does not have to do that anymore