replace GitRunner-usage with run-context
This commit replaces the use of `GitRunner`
in `datalad_next.constraints.git` with a
use of `datalad_next.runners.run.run`-context.

Remark: `GitRunner` is still used internally
to provide an adjusted environment for the
execution of `git` in the `run`-context.

With this commit all subprocess executions
in `datalad-next` are based on the run-context.