add batch-context managers and tests
Co-authored-by: Michael Hanke <michael.hanke@gmail.com>

This commit adds a `batchcommand`-context-manager that
keeps track of a subprocess and supports batch-style
communication. That means, it allows to send a request
to the subprocess and will return the response from the
subprocess. How the output of the subprocess has to be
interpreted to yield a response is defined by the
protocol that is provided to the
batchcommand-context-manager. The
batchcommand-context-manager uses the run-context-manager
to ensure that the exit-handler waits for process
termination.

The commit also contains three convenience functions
that provide context-manager for common interaction
patterns (common in `datalad` and `datalad_next` anyways):

1. `stdout_batchcommand`: this context-manager writes to
   `stdin` of the subprocess and returns the next, yet
   unread, chunk of bytes that was written to `stdout`
   of the subprocess.

2. `stdoutline_batchcommand`: this context-manager writes
   to `stdin` of the subprocess and returns the next,
   yet unread, line that was written to `stdout` of the
   subprocess. Line-separators and whether they are
   included in the output can be chosen when creating
   the context-manager.

3. `annexjson_batchcommand`: this context-mananger writes
   to `stdin` of the subprocess and returns a JSON-object,
   that is created from the next, yet unread, line that
   is written to `stdout` of the subprocess. This
   context-mamanger is geared towards git-annex batch
   commands with JSON-result output, i.e. git-annex
   invocations with the parameters `--batch`, `--json`,
   `--json-error-messages`.

The tests provide an example for protocol that handles
output that varies in size, i.e. output of
python-programs. The class that implements the protocol
is called `PythonProtocol`.