warn if an exception leaves the run-context
This commit adds a warning that describes the possibility
of a stall, if an exception leaves the run-context.

The reason for the possible stall is that it is not clear
whether the code in the run-context did trigger a
subprocess exit. If it did not, and if no timeouts are
set, the exit handler of the run-context might wait
indefinitely for the subprocess to exit.

If the code ensures that all exceptions are caught, e.g.
by using a try-finally clause, and the subprocess exit
is triggered in case of an exception, the stalling
can be prevented.