New `datalad_next.runners.iter_subproc`
This is a convenience wrapper around `datalad_next.iterable_subprocess`.
The main purpose is to have a place for documentation and feature
additions -- without adding needless diff to the original implementation
that was merged in from an external source.

The wrapper uses `COPY_BUFSIZE` as default chunk size (which is
numerically identical to the hardcoded default on POSIX systems).
This should be better than a separate literal.

A second different is the option to run a subprocess without any
explicit input.