Replace regex approach to ls-files output parsing
In #540 I found about ~10% of the runtime to be attributable to the
regex-matching. This patch replaces the regex with two split()
calls. In local benchmarks I see a ~10% speedup.

One major other source of slow-down was/is the construction of
`PurePosixPath` objects. It seems this is being taken care of by
https://github.com/python/cpython/issues/101362 and will be resolved
eventually.

Closes #540