Remove default separator from `itemize()`
This addresses an issues brought up in
https://github.com/datalad/datalad-next/pull/539#discussion_r1403427421

This changeset removes the default argument to avoid the impression that
"line-processing" is the main target. The code does not imply that,
and the existing usage also not.

The possibility to do line-splitting is not touched (or removed). The
documentation needs no adjustment.