improve `iter_annexworktree`
This commit does the following:

1. Remove use of `join_with_list` from `iter_annexworktree`

Instead of building a list, `AnnexWorktree*`-objects are
created immediately, and ammended during processing. This
is based on an idea in reviewer comment:
https://github.com/datalad/datalad-next/pull/539#discussion_r1408816818

2. add support for `AnnexWorktreeFileSystemItem`

If the underlying iter_gitworktree-iterator generates a
`GitWorktreeFileSystemItem`, this version of `iter_annexworktree`
will create an `AnnexWorktreeFileSystemItem`. This supports
the `fd`-argument.

This addresses reviewer comment:
https://github.com/datalad/datalad-next/pull/539#pullrequestreview-1754488046

3. Adds comments and renames variables to clarify code

This addresses reviewer comments:
https://github.com/datalad/datalad-next/pull/539#discussion_r1408811577