use built-in line endings in `iter_annexworktree`
This commit set the `sep`-argument in calls to
`datalad_next.itertools.itemize` to `None`. This
triggers the code to use the built-in line endings,
which include numerous line endings, from
different OSs.

This is done to take care of changing behavior
in underlying processes, in this case in
`git-annex`.

This commit address the reviewer-comment:
https://github.com/datalad/datalad-next/pull/539#discussion_r1410188112