simplify `itemize`-ing of `git annex find`-ouput
Thie commit uses a `b'\n'` separator when itemizing
the output of `git annex find` and does not keep
line endings. This simplifies the call to
`itemize` and the test for a non-empty key in
the splitter-function of the enclosing
`route_out`.

It also requires to add `b'\n'` to drive the
consuming `git annex examinekey`-subprocess.
This is done with `intersperse`.