use documented newline-encoding in git annex format
This commit fixes a format string in a git annex find
command. The documentation states that a ``\n´´, i.e.
two characters: a backslash and an ``n´´, instructs
git annex to write a newline-character.
We have used the python string '\n', i.e. a single
character: newline. Although git annex seems to
accept newline and emits a newline, that is
undocumented behavior and should therefore not be
used.