remove `align_pattern` from the ssh-stream
This commit updates the docstring of `align_pattern`
to document that `align_pattern` can be "removed"
from a stream after it yields an item.

This is used in `SshUrlOperations._get_props`
to return a "clean" stream-iterator, i.e. a stream
iterator that does not still execute the
`align_pattern`-code.

(The code was actually already doing the right
thing, but comments were still suggesting, that
`align_pattern` stays in the loop.)