ensure `IterableSubprocessError` generation
This commit catches `StopIteration`-excptions
that are raised in an `iter_subproc`-context
to ensure that `IterableSubprocessError`
generation is performed (if the subprocess
returned non-zero).

Without catching `StopIteration`-exceptions
within the subproc-context, a `next()` on an
exhausted iterator would trigger a code-path
that does not check the
return code of the subprocess. This path is
actually triggered by almost all exceptions,
but we generate `StopIteration` as part of
our error-free program flow, so we have to
deal with it at some point.

There are a few possible solutions here:

1. Do what this commit does. Don't let
   expected exceptions leave the
   subproc-context.

2. Modify the code in such a way, that
   it does not use `next` on iterators.

3. Ensure that `iterable_subprocess`
   executes the check for
   `IterableSubprocessError` in the
   presence of an exception from the
   context. This could easily be done
   by putting the check into a
   `finally`-branch. But that would
   "hide" a code exception, if the
   subprocess does exit with non-zero.
   (And would not pass a lot of tests)