handle `StopIteration` in `iterable_subprocess`
This commit modifies `iterable_subprocess` so that
it prioritizes `IterableSubprocessError` exceptions
over `StopIteration` exceptions, if a `StopIteration`
exception is raised in the body of an
`iterable_subprocess`-context and if the sub process
exits with a non-zero return code.

`StopIteration` in conjunction with a non-zero
return code indicates an unexpected short read
due to an early "error"-exit of the sub process.

This commit is a result of this discussion:
https://github.com/datalad/datalad-next/pull/546#discussion_r1418976060