convert generator-function into generator
This commit converts the generator function
`output_from` into an instance of a the
generator class `OutputFrom`.

The rationale for this conversion is that
the generator class can store the
return code of the subprocess as attribute.
Thus the returncode is available, even
if an exception is raised within the
`iterable_subprocess`-context.