use `returncode` of iterable_subproccess generator
This commit modifies SshUrlOperations to make use
of the `returncode`-attribute of the generator that
is returned by `with iterable_subprocess`.

This allows to fetch the return code of the subprocess
even if no `IterableSubprocessError` was raised.
This situation can occur, if an exception is raised
in an `iterable_subprocess`-context.