fix `props` accessed before assignment
In some cases ssh-commands return a zero
exit value, but we cannot read all expected
data, i.e. we get a `StopIteration` exception
when calling `next(stream)`. This was not
expected and lead to an access to a not
existing variable.

This commit fixes this by assuming that
the combination of a `0` return code and
a short read indicates that the file in
question does not exist anymore.