handle `StopIteration` exceptions separately
In some cases `SshUrlOperations._get_props()`
raises a `StopIteration`-exceptions. This
usually indicates a non-existing resource.
The return code of the `ssh`-subprocess might
be 255, because the process got terminated
when leaving the `iter_subproc` context (due
to the exception) before it exited with 244,
which would usually indicate a non-existing
resource.

We therefore handle `StopIteration` differntly,
i.e. we do not check the return code if a
`StopIteration`-exception occurs. Instead
we always assume a non-existing resource.