Add align_pattern iterator and tests
This feature eases the processing of output that is segmented
using "magic-markers", e.g. as done in `SshUrlOperation` for
separating the stats on a download from the actual payload.

This commit prepares for a change in that implementation that
migrates the SSH client execution to an `iter_subproc` based
implementation.