Add an iterator that reports stream progress
The iterator is made available as an internal
`_with_progress()` method of `UrlOperations`.

It wraps a stream (or more generally an iterator),
and calls the existing helpers for progress reporting
as a "side-effect".

Progress end reporting is done in a `finally` branch
by `side_effect`, hence is robust to exceptions
during stream processing.