Replace `GitRunner` usage with plain `subprocess`
This is a step towards #541. While we may want to have an abstraction
layer for such simple calls, there is no point in keeping the heavy
`GitRunner` here.

Possible a trivial wrapper around `subprocess.run()`, further triming
the space of possibilities would be useful.

But this PR is not the space for a big-picture change.