Enhance `iter_subproc` with a `cwd` parameter
Running in a particular working directory is a very common paradigm
(think in a Git repository). So far, we used Git's `-C` option
to perform the directory changes. This is fine, fast, and works
universally.

However, recent discussion asked whether it would make sense to
call annex commands via `git-annex`, rather than `git annex`.
The former does not provide `-C`. This change readies the code
for such a move.

Another use case for this is the passing of (many) file paths.
With an option to run in a particular directory, the file names
can be just that (not full paths). This can help to mitigate
situations, where many file paths potentially exhaust the max
command length.