prevent exceptions from leaving subproc-context
This commit modifies SshUrlOperations so that it
catches expected exceptions before they leave the
`iter_subproc`-context. This leads to a proper
shutdown of the ssh-subprocess, via closing its
stdin, and to a reliable signaling of the
return code of the remote command.

The latter would not always be the case if the
`iter_subproc`-context is left with an
exception, because that triggers a `terminate()`
on the shell process, and the exit code of
the shell process will reflect this termination.