Uplift `create_sibling_webdav` parameter validation
Use `EnsureRemoteName` for sibling and storage sibling parameters.

Now immediately checks for conflicting remote name, when
`existing='error'` and not going recursive. In this case,
the command would error later, and in a more expensive way.

Now this situation is caught during command validation, and
reported properly, and jointly with others.