Implement `iter_gittree()` and support it in `ls_file_collections`
This is the first iterator that is not simply reporting on some file
system that is identifiable via a single location parameter.  Instead,
it requires a repository location and a second "tree-ish" identifier.

I decided to skip the implementation of a file-pointer inclusion
into the returned items. Two reasons: (1) to make this efficient
it would need a lazy-execution of `git cat-file`, and (2) I cannot
think of a use case for content hashing beyond identification,
and that use case is perfectly addressed by the reported `gitsha`
already.

For the integration with `ls-file-collection` I decided to stay
close to the behavior of `git ls-tree` itself. The collection identifier
parameter is used for the tree-ish, while the repository selection is
done via the working directory. This includes constraining reports
to subdirectories (in non-bare repos), like `git ls-tree` does too.

Closes: #349