Change semantics of `PathBasedItem` names
Previously, the `name` attribute was some kind of `PurePath` instance.
The precise nature varied across item types and corresponding
iterators.

This led to two problems:

- Performance (see #581): the creation of a Path instance is not
  particularly cheap. If a consumer does not need such an instance,
  and unconditional creation is a waste of resources
- Abstraction (see #430): the Path representation of a path-like
  name does not always match the exact semantics of the original
  name (it might have a trailing slash that has a purpose, etc)

Both problems are now addressed by relaxing the type of `.name`
of a `PathBasedItem`. It can now be anything. Consumers that require
an actual Path instance can use the `.path` attribute. An analog
access is also implemented for `.link_target` (which now remains
literal), that is accompanied by `.link_target_path`, where it
was necessary.

This change in approach removed the need for the `_ZipFileDirPath`
workaround that was used to re-establish compatibility of a Path-based
`.name` with the conventions in a `ZipFile` collection/container.
With the `.name` attribute retaining its native semantics, the
workaround is no longer needed as was removed.

In order to document the (lack of) implemented homogeneous conventions
for path-based items, the docstrings of all iterators have been amended
with information on the nature of the `.name` attribute yielded by them.
The corresponding data classes have received docstrings for the newly
added properties for Path access.

These properties uniformly use the `cached_property` decorator.
For the lifetime of an item, the nature of such a path should never
change, and caching it automatically addresses the significant
creation cost on accessing a path representation repeatedly.

Closes #554
Closes #581