Common helper to call Git subprocesses
This is a thin wrapper around `iter_subproc` that is to be used
unconditionally whenever the command given to `iter_subproc` would
start with `git`.

Presently, this wrapper does nothing but adding that `'git'` argument
to the front of any argument list.

Eventually, it might be used to call a configurable Git binary,
establish certain environment conditions for all Git calls, etc.

Right now, this is just cheap future-proofing.

Closes #584