feat: new support subpackage for Git's pathspecs
The main (if not only) purpose of this functionality is pathspec
mangling/translation for handing them over to analog Git command
calls on submodules -- for any Git command that supports pathspecs,
but not recursion.

A simple example for such a command is `git ls-files --other`. It
accepts pathspecs, but does not implement `--recurse-submodules` for
listing untracked files.

The goal of this functionality is to be able to take pathspecs that is
valid in the context of a top-level repository, and translate it such
that the set of paths specs given to the same command running on/in a
submodule/subdirectory gives the same results, as if the initial
top-level invocation reported them (if it even could).

The key algorihtm lives in a standlone function
`yield_subdir_match_remainder_pathspecs()` that performs a purely
lexical analysis. It also comes with a dedicated test collection that is
leaner and easier to extend than the previous one (which remains also).

The additionally included sketch of a testbattery uses ``git ls-files --other`
for testing, rather than a formal description -- because the behavior
of the implementation is more elaborate than the documentation at
https://git-scm.com/docs/gitglossary#Documentation/gitglossary.txt-aiddefpathspecapathspec
suggests.

All testing is (for now) performed within a single repository, and with
translation for execution in subdirectories.

Refs: #587, https://github.com/datalad/datalad/issues/6933