Add submodule recursion support to `iter_gitdiff()`
This also includes the ability to report tree item. This is an
orthogonal switch (analog to `diff-tree -t`) that also supports
submodule-type items as subtree variants.

This is different/better than the handling in datalad-core, where such
items (submodule only) are reported unconditionally, and need to be
filtered out, if the container modification is not interesting
(ie., when working with a hierarchy monorepo-style).