move `CommandError` to `datalad_next.exceptions`
This commit removes the definition of `CommandError`
in `datalad_next.runners.exceptions`. Instead it
imports and tests the patch `CommandError` from
`datalad.runner.exception` into
`datalad_next.exceptions`.

The commit implements more or less point 1. of comment:
https://github.com/datalad/datalad-next/pull/579#pullrequestreview-1805653383
The difference is that there is no more definition
of `CommandError` in `datalad_next`, just a patch
of `datalad.runner.exception.CommandError`.