Initial draft of a dataclass as basis for DataLad command results
This implements the plan outlined at
https://github.com/datalad/datalad-next/issues/418, albeit in an even
leaner way.

No additional patches of datalad-core code were necessary.

The included `CommandResult` class is intended to be a base class for
further command-specific result types. A demo for that will follow
in a subsequent commit.

Unlike a classic "result-dict", property values are encouraged to
be specialized types (not strings), to facilitate further processing.
This includes in particular instances of `enum`, `Dataset`, and `Path`
values.

In order to enable a smooth transition, this class implements minimal
aspects of the `dict` API. Enough to be compatible with the processing
by the main datalad-core result loop. This enables such a dataclass
instance to pass-through the result loop without a type-change. This
aspects enabled further reuse and reliance on a command-specific result
type, for example, by  a command's tailored result handler.

The implemented `dict` API is also sufficient for JSON-serialization to
continue to work, as implemented in the generic result renderers.

This approach comes with a cost for as long as we want to support this
legacy-compatibility: no result can ever have any of the following
properties (`get`, `pop`, `items`).

Closes #418