Homogenize ls-files property handling in `iter_gitworktree()`
Previously, properties were often, but not always passed on to the
pending item. This made the implementation susceptible to complex
interactions based on the order and status of items discovered
in a repo, especially when withholding items in recursion='no'
operation.

This change unconditionally passes any known properties onto the
pending item, and ensures that any decision making is always aware
of the complete item description.