Add `GitDiffItem.modification_types`
`GitDiffItem` can now carry additional modification qualifiers. This
enables reports similar to `git status` that can tell whether a
submodule has new commits, modified content, or untracked content --
which are all variants of a container-item modification, and
can co-occur for the same item.

We support them for any item, in order to also be able to express
such modification qualifiers for directory-type containers.