Make `iter_git|annexworktree()` robust to deleted files
When reporting file pointers. Both stumbled over files that were still
reported by `ls-file`, but did not exist on the filesystem anymore
(this happens when a deletion is not commited/staged).

This change makes both iterators fall back on yielding a non-filesystem
work tree item in this case.

The corresponding tests have been adjusted to cause and adapt to this
case.

Closes #582