robustify tests of `iterable_subprocess`
This commit modifies the test
`test_success_returncode_available_from_generator_with_exception`
to accept `0` and `-15` return codes. The reason is that
any `StopIteration`-exception that occurs before the
subprocess has exited will lead to a subprocess termination
and to a `-15` return code.
There seem to be system configurations in which this
situation appears.

The new test is renamed to
`test_returncode_available_from_generator_with_exception`.