Let DataLad's log level adjust the datalad-annex verbosity threshold
Previously, one had to call `git-clone/fetch` directly with some
`-v` options in order to trigger verbose output of the remote helper.

With this change a DataLad log level of 'debug' (or more verbose),
also triggers verbose output in the helper. This helps with
accessibility of this feature for internal Git calls.

Note that due to a limitation of the `datalad -l` CLI option handling
the log-level is not propagated to subprocesses. This only works
for env specification ala DATALAD_LOG_LEVEL=debug