Test for a repository deposit before establishing mirrorrepo
The `mirrorrepo` property of the datalad-annex Git remote helper
implementation creates an empty repository, if the remote has no
refs (or does not exist). This is sensible behavior for enabling
initial pushes.

However, it leads to successful but empty clones/fetches for any
misspecified URL or connectivity failure. This is undesirable.

This change adds an explicit test for an available/accessible
repository deposit, whenever a clone/fetch is performed. Wrong URL
specifications or connectivity issues will now lead to an explicit
error exit. This fact is communicated via a non-verbose message,
and the full error is available at verbosity level 2, i.e. with a
`git clone -v`.

Closes: gh-636