add `bufsize` parameter to iterable subprocesses
This commit adds the parameter `bufsize` to
`iterable_subprocesses` and to `iter_subproc`.
Its value is passed into `subprocess.Popen`.

This change is necessary to implement
interactions with interactive subprocesses,
for example a remote shell. These
processes usually require a buffer size
of zero, because each "request", i.e. an
input to an interactive process, results
in a "response" of unknown size and
structure. In an interactive communication,
each repsonse has to be interpreted before
the next request can be sent. Therefore
every byte that an interactive process emits
has to be sent to the client, which
processes it and assembles a response.

In addition this commit makes the previously
internal class `OutputFrom` (defined in
`datalad_next.iterable_subprocess.iterable_subprocess`)
available as
`datalad_next.iterable_subprocess.OutputFrom`.

This commit is in preparation for an
interactive shell-class that will be
used, for example, in RIA stores.